# src/main/java/cvp/apiProduto/dto/ProdutoResponseDTO.java
package cvp.apiProduto.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class ProdutoResponseDTO {

    private Long id;
    private String nome;
    private BigDecimal preco;
}